impl2/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
aiofiles==23.2.1
    # via semantic-kernel
aiohttp==3.9.5
    # via semantic-kernel
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
    #   starlette
asgiref==3.8.1
    # via openapi-core
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
black==24.1.1
    # via -r .\requirements.in
certifi==2024.2.2
    # via
    #   -r .\requirements.in
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via prance
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
coverage[toml]==7.5.0
    # via pytest-cov
defusedxml==0.7.1
    # via semantic-kernel
distro==1.9.0
    # via openai
dnspython==2.6.1
    # via
    #   -r .\requirements.in
    #   pymongo
docopt==0.6.2
    # via -r .\requirements.in
fastapi==0.110.2
    # via -r .\requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   -r .\requirements.in
    #   openai
hypercorn==0.16.0
    # via -r .\requirements.in
hyperframe==6.0.1
    # via h2
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
isodate==0.6.1
    # via
    #   openapi-core
    #   rdflib
jinja2==3.1.3
    # via -r .\requirements.in
jsonschema==4.21.1
    # via
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.2
    # via openapi-spec-validator
jsonschema-spec==0.2.4
    # via openapi-core
jsonschema-specifications==2023.7.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.10.0
    # via openapi-spec-validator
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
more-itertools==10.2.0
    # via openapi-core
motor==3.4.0
    # via semantic-kernel
multidict==6.0.5
    # via
    #   -r .\requirements.in
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
numpy==1.26.4
    # via semantic-kernel
openai==1.23.6
    # via
    #   -r .\requirements.in
    #   semantic-kernel
openapi-core==0.18.2
    # via semantic-kernel
openapi-schema-validator==0.6.2
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.1
    # via openapi-core
packaging==23.2
    # via
    #   black
    #   prance
    #   pytest
parse==1.20.1
    # via openapi-core
pathable==0.4.3
    # via
    #   jsonschema-path
    #   jsonschema-spec
pathspec==0.12.1
    # via black
platformdirs==4.2.0
    # via black
pluggy==1.5.0
    # via pytest
prance==23.6.21.0
    # via semantic-kernel
priority==2.0.0
    # via hypercorn
psutil==5.9.8
    # via -r .\requirements.in
pydantic==2.7.1
    # via
    #   fastapi
    #   openai
    #   semantic-kernel
pydantic-core==2.18.2
    # via pydantic
pymongo==4.7.0
    # via
    #   -r .\requirements.in
    #   motor
pyparsing==3.1.1
    # via rdflib
pytest==7.3.2
    # via
    #   -r .\requirements.in
    #   pytest-cov
pytest-cov==5.0.0
    # via -r .\requirements.in
python-dotenv==1.0.1
    # via
    #   -r .\requirements.in
    #   semantic-kernel
python-multipart==0.0.9
    # via -r .\requirements.in
pyyaml==6.0.1
    # via
    #   jsonschema-path
    #   jsonschema-spec
rdflib==7.0.0
    # via -r .\requirements.in
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-spec
    #   jsonschema-specifications
regex==2023.12.25
    # via
    #   semantic-kernel
    #   tiktoken
requests==2.31.0
    # via
    #   -r .\requirements.in
    #   jsonschema-path
    #   jsonschema-spec
    #   prance
    #   tiktoken
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.6
    # via prance
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
semantic-kernel==0.9.1b1
    # via -r .\requirements.in
six==1.16.0
    # via
    #   isodate
    #   prance
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.37.2
    # via fastapi
tiktoken==0.6.0
    # via -r .\requirements.in
tqdm==4.66.2
    # via openai
typing-extensions==4.11.0
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
uvicorn==0.29.0
    # via -r .\requirements.in
werkzeug==3.0.2
    # via openapi-core
wsproto==1.2.0
    # via hypercorn
xmlformatter==0.2.6
    # via -r .\requirements.in
yarl==1.9.4
    # via aiohttp
